Revision codigo desde main y archivos y funciones no utilizadas
